Repo rescue down under, part 4
Okay, this is an important one. Utility functions are renamed, thermal curve figure is migrated, but most importantly, a bunch of broken scripts for data extraction have been consigned to oblivion. They wrote out temp files that have also been locally deleted. They have been replaced with the correct version of the script (which ended up in the Dropbox, unmigrated) and the file it writes out is staying in the Dropbox for now.